tests: rename default plan to build-gating to better eflect its current purpose
My expectation is that in the future, more tests will land in this git,
and not all of them would be useful in gating time. AFAIK, CI systems do
run everything they can, therefore adding `when` to control when the
plan is executed - `is not defined` bit should allow anyone to run it
manually, while the rest of the condition should limit its presence to
gating events.