Remove _smp_mflags override on ppc64le
Using a fixed -j argument is pretty dubious and may have the
opposite of the desired effect. E.g. on a recent copr build I
see ppc64le having 5 cpus, so this actually increases the level
of parallelism (and oversubscribes...) rather than decreasing.

I don't think we have evidence that this particular workaround is
still needed, so drop it.