Skip openmp tests when builder uses 5-level page tables
A large number of openmp tests using tsan fail when we hit certain
machines on the rhel8-beefy channel in brew, because they appear
to the use 5-level page tables. This results in memory being
mapped in places where tsan does not expect it.
See https://github.com/llvm/llvm-project/issues/111492 for more
context.

Work around this by disabling the openmp tests that use tsan if
the cpu has the la57 feature.