Port changes from RHEL 9
This change brings back some patches we had applied in LLVM 18.
And since the `bundle_compat_lib` switch in RHEL still builds LLVM 18,
I've added them here. This was easily possible due to #323.

This effectively allows us to build LLVM 19 in RHEL9 (see also RHEL-57461).

I've also added the `--gcc-install-dir` to the config file which is used
once clang is installed. This is to tell clang in RHEL which standard
library to link against.

We decided to no longer patch clang to default to DWARF4. Instead we tune
the default by adding `-gdwarf-4` to the config file.

RHEL-wise we've bumped the gts version from 13 to 14 (see RHEL-38228).