[make] Use mock instead of rpmbuild
I always wanted to use `mock` instead of `rpmbuild` for building release
and snapshot builds. I disliked `mock` for a very particular reason.
When a build failed and I wanted to go in and change the last running
script and re-run it, I thought this wasn't possible. Little did I know.

Here are the make targets and what they do now:

```
Available targets
-----------------
get-sources-snapshot                  - Downloads all sources we need
for a snapshot build.
get-sources-release                   - Downloads all sources we need
for a release build.
srpm-release                          - Builds an SRPM that can be used
for a release build.
srpm-snapshot                         - Builds an SRPM that can be used
for a snapshot build.
scrub-chroot                          - Completely remove the fedora
chroot and cache.
mockbuild-release                     - Start a mock build of the
release SRPM.
mockbuild-snapshot                    - Start a mock build of the
snapshot SRPM.
edit-last-failing-script              - Opens the last failing or
running script from mock in your editor
                                        of choice for you to edit it and
later re-run it in mock with:
                                        "make
mockbuild-rerun-last-script-...".
mockbuild-rerun-last-script           - Re-runs the last failing or
running script of your release/mock mockbuild.
help                                  - Display this help text.
get-llvm-version-release              - Determines the LLVM version
given in the llvm.spec file.
get-llvm-version-snapshot             - Determines the LLVM version
given in the version.spec.inc file.
get-spec-file-release                 - Parses the spec file for the
Release: tag
get-srpm-release                      - Determines the name of the SRPM
used for release builds
                                        Can be overriden by giving "make
... SRPM_PATH=foo.src.rpm".
get-srpm-snapshot                     - Determines the name of the SRPM
used for snapshot builds
                                        Can be overriden by giving "make
... SRPM_PATH=foo.src.rpm".
```

When you want to build a release build for fedora you can do so by
running `make mockbuild-release` it will download the sources for you
and create an SRPM that it will pass to `mock` for the final build.

To build for `centos` you can use `make mockbuild-release
MOCK_CHROOT=centos-stream-9-x86_64`.