Use new config already for current build
offload-tblgen currently fails to build in the rhel-9 chroot for
the following reason: The host compiler uses gcc-toolset-13 via
compiled-in option. The newly built compiler uses gcc-toolset-14
via config file. However, because the config file is not installed
yet, when we use this compiler during the runtimes build, we use
system libstdc++. This results in a mismatch in the libstdc++
version between LLVMSupport and offload-tblgen.

I'm fixing this by already explicitly using the new config file
via optflags. This makes sure that the runtimes and non-runtimes
builds use the same configuration, and should also prevent issues
when clang can no longer bootstrap with the new config.