Disable libcxx conditional on Fedora <= 40
Fedora 41 was the first version that enabled FatLTO for clang-built
files. FatLTO is important in order to link binaries with GNU LD, which
is the linker used while building the LLVM runtimes.

Considering the libcxx conditional has been added recently in rawhide
and will only be used in Fedora >= 42 and RHEL versions, I believe it's
safe to simply disable it when building on Fedora <= 40, considering
this setup will only be used while testing LLVM Snapshots for a few
months until this Fedora version reaches EOL.