Do not create symlinks on i686
Currently, it's not possible to install llvm-devel.i686 and
llvm-devel.x86_64 at the same time, breaking multilib support.
The reason is that binary symlinks between the two packages
conflict, and unlikely for file conflicts, dnf treats these as
an error.

Fix this by not shipping any symlinks in the default prefix on
i686. This means that i686 packages only ship the contents of
/usr/lib/llvmN. This allows conflict-free co-installation with
the x86_64 packages.

I have opted to drop *all* of the symlinks in the default prefix
on i686, but this is not strictly necessary. E.g. keeping lib
symlinks should be fine. But I think it's cleaner to just say
that i686 has everything in the llvm prefix only, rather than
trying to be more precise.