Use libdir suffix in versioned prefix as well
Use the lib/lib64 distinction for the versioned prefix, to avoid
a mismatch with the default prefix. This means that the versioned
prefix now has libraries in /usr/lib64/llvmN/lib64.

This should fix the issue https://src.fedoraproject.org/rpms/llvm/pull-request/446
ran into. This would also be needed if we wanted to switch to
installing into libexec.