Backport rust related fixes for ppc64le
Resolve rhbz#2382683, the fix is not ppc specific, but does fix the
miscompile when building a pgo optimized rust toolchain.

Also, apply a fix for a bug observed in c9s/c10s caused by a miscompile
when spilling crbits when rust is built with power9 optimizations. This
fix is informally approved, but still pending.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2382683
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2383037