Opt-out of debuginfo for static libraries
Since https://fedoraproject.org/wiki/Changes/StaticLibraryPreserveDebuginfo
debuginfo in static libraries is preserved by default. Opt out of
this for the LLVM package.

Debuginfo for static libraries in llvm-static is huge: The RPM
went from ~45MB to ~450MB with this change. The uncompressed size
is multiple GB. llvm-static is a dependency of llvm-devel, so
anyone using that package has to pay for this, even if they are
not using static libraries (which they don't by default).