Specify lit --timeout
If a lit test hangs, make sure our build is not going to hang
forever. It seems like this happened with s390x/ppc64le jobs on
a recent snapshot build, which makes it impossible to see what
the actually hanging test was, and also wastes a lot of resources.

[skip changelog]