Explicitly run rpmautospec on copr
As we use the make_srpm (rather than dist-git) method, we need to
explicitly run rpmautospec in the .copr/Makefile.

On distros other than centos-stream-9 we get away with not doing
this because rpmautospec is available during the main build and
we can use fallback expansions. But on centos-stream-9 it's
important that this happens during the make_srpm phase.

[skip changelog]