[make] prepare-copr target
This target lets you run `make prepare-copr IP=<COPR_IP>` where `<COPR_IP>`
is the IP address of a copr host to which you can SSH.

`prepare-copr` rsyncs the current directory (excluding `.git/`)
to the copr instance specified with `<COPR_IP>` and then runs the
`prepare-copr.sh` script on the instance. This ensures you have a good
experience with tools like `tmux`, `vim`, and `make` installed for you
as well as mock directories set up.

Also, we make an attempt to prolong the instance time by 24 hours.

The old and now removed `prepare-copr-instance` target and its
dependents made it harder than it needs to be to set up a copr instance
for inspection.

[skip changelog]