Let LLD and LTO get used independently of PGO
I suspect the lines that enable LLD usage got inside of the %{with pgo}
block by mistake. The LTO settings must be used whenever lto_build is enabled
and is independent of enabling pgo. In order to avoid enabling LTO on
RHEL 8, explicitly ask to disable this scenario.

Add the conditional use_lld that will allow to control when LLVM should
be linked using LLD. This will allow to use LLD even when PGO is
disabled. This is enabled by default, except for i686 (all distros),
Fedora <= 42 and RHEL 8.

This is a NFC.

[skip changelog]