Treat the case when GTS version used is different from the expected
In rare cases, the newly built clang may end using a GTS version that is
not equal to the expected one.
One of the reasons on why this happens is because the system clang may
depend on a older GTS version while we want to use a newer one.

In that scenario, we install a libatomic from the latest GTS, but the
system clang sets a config file in /etc/clang pointing to the older
clang. Until that file is replaced, the newly built clang will end up
using the older GCC directory which doesn't have libatomic, causing test
failures.

This is a temporary scenario that is automatically fixed when the newly
built clang is packaged and is installed on the system.